+++ readium/navigator/src/main/java/org/readium/r2/navigator/R2BasicWebView.kt
@@ -53,6 +53,9 @@ open class R2BasicWebView(context: Context, attrs: AttributeSet) : WebView(conte
         fun onPageEnded(end: Boolean)
         fun onScroll()
         fun onTap(point: PointF): Boolean
+        fun onDragStart(event: DragEvent): Boolean
+        fun onDragMove(event: DragEvent): Boolean
+        fun onDragEnd(event: DragEvent): Boolean
         fun onDecorationActivated(id: DecorationId, group: String, rect: RectF, point: PointF): Boolean = false
         fun onProgressionChanged()
         fun onHighlightActivated(id: String)
@@ -72,7 +75,6 @@ open class R2BasicWebView(context: Context, attrs: AttributeSet) : WebView(conte
     }
 
     lateinit var listener: Listener
-    lateinit var navigator: Navigator
     internal var preferences: SharedPreferences? = null
 
     var resourceUrl: String? = null
@@ -361,6 +363,71 @@ open class R2BasicWebView(context: Context, attrs: AttributeSet) : WebView(conte
         mPopupWindow.showAtLocation(this, Gravity.CENTER, 0, 0)
 
         return true
+    }
+
+    @android.webkit.JavascriptInterface
+    fun onDragStart(eventJson: String): Boolean {
+        val event = DragEvent.fromJSON(eventJson)?.takeIf { it.isValid }
+            ?: return false
+
+        return runBlocking(uiScope.coroutineContext) { listener.onDragStart(event) }
+    }
+
+    @android.webkit.JavascriptInterface
+    fun onDragMove(eventJson: String): Boolean {
+        val event = DragEvent.fromJSON(eventJson)?.takeIf { it.isValid }
+            ?: return false
+
+        return runBlocking(uiScope.coroutineContext) { listener.onDragMove(event) }
+    }
+
+    @android.webkit.JavascriptInterface
+    fun onDragEnd(eventJson: String): Boolean {
+        val event = DragEvent.fromJSON(eventJson)?.takeIf { it.isValid }
+            ?: return false
+
+        return runBlocking(uiScope.coroutineContext) { listener.onDragEnd(event) }
+    }
+
+    /** Produced by gestures.js */
+    data class DragEvent(
+        val defaultPrevented: Boolean,
+        val startPoint: PointF,
+        val currentPoint: PointF,
+        val offset: PointF,
+        val interactiveElement: String?
+    ) {
+        internal val isValid: Boolean get() =
+            !defaultPrevented && (interactiveElement == null)
+
+        companion object {
+            fun fromJSONObject(obj: JSONObject?): DragEvent? {
+                obj ?: return null
+
+                val x = obj.optDouble("x").toFloat()
+                val y = obj.optDouble("y").toFloat()
+
+                return DragEvent(
+                    defaultPrevented = obj.optBoolean("defaultPrevented"),
+                    startPoint = PointF(
+                        obj.optDouble("startX").toFloat(),
+                        obj.optDouble("startY").toFloat()
+                    ),
+                    currentPoint = PointF(
+                        obj.optDouble("currentX").toFloat(),
+                        obj.optDouble("currentY").toFloat()
+                    ),
+                    offset = PointF(
+                        obj.optDouble("offsetX").toFloat(),
+                        obj.optDouble("offsetY").toFloat()
+                    ),
+                    interactiveElement = obj.optNullableString("interactiveElement")
+                )
+            }
+
+            fun fromJSON(json: String): DragEvent? =
+                fromJSONObject(tryOrNull { JSONObject(json) })
+        }
     }
 
     @android.webkit.JavascriptInterface