+++ readium/navigator/src/main/java/org/readium/r2/navigator/epub/EpubNavigatorFragment.kt
@@ -85,6 +85,11 @@ class EpubNavigatorFragment private constructor(
          * Provide one if you want to customize the selection context menu items.
          */
         var selectionActionModeCallback: ActionMode.Callback? = null,
+
+        /**
+         * Whether padding accounting for display cutouts should be applied.
+         */
+        val shouldApplyInsetsPadding: Boolean? = true,
     )
 
     interface PaginationListener {
@@ -367,7 +372,7 @@ class EpubNavigatorFragment private constructor(
                 ?: return null
 
         val rect = json.optRectF("rect")
-            ?.apply { adjustToViewport() }
+            ?.run { adjustedToViewport() }
 
         return Selection(
             locator = currentLocator.value.copy(
@@ -381,17 +386,15 @@ class EpubNavigatorFragment private constructor(
         run(viewModel.clearSelection())
     }
 
-    private fun PointF.adjustToViewport() {
+    private fun PointF.adjustedToViewport(): PointF =
         currentFragment?.paddingTop?.let { top ->
-            y += top
-        }
-    }
+            PointF(x, y + top)
+        } ?: this
 
-    private fun RectF.adjustToViewport() {
-        currentFragment?.paddingTop?.let { top ->
-            this.top += top
-        }
-    }
+    private fun RectF.adjustedToViewport(): RectF =
+        currentFragment?.paddingTop?.let { topOffset ->
+            RectF(left, top + topOffset, right, bottom)
+        } ?: this
 
     // DecorableNavigator
 
@@ -414,6 +417,7 @@ class EpubNavigatorFragment private constructor(
 
     internal val webViewListener: R2BasicWebView.Listener = WebViewListener()
 
+    @OptIn(ExperimentalDragGesture::class)
     private inner class WebViewListener : R2BasicWebView.Listener {
 
         override val readingProgression: ReadingProgression
@@ -431,22 +435,6 @@ class EpubNavigatorFragment private constructor(
 
         override fun onPageChanged(pageIndex: Int, totalPages: Int, url: String) {
             r2Activity?.onPageChanged(pageIndex = pageIndex, totalPages = totalPages, url = url)
-            if(paginationListener != null) {
-                // Find current locator
-                val resource = publication.readingOrder[resourcePager.currentItem]
-                val progression = currentFragment?.webView?.progression?.coerceIn(0.0, 1.0) ?: 0.0
-                val positions = publication.positionsByResource[resource.href]?.takeIf { it.isNotEmpty() }
-                    ?: return
-
-                val positionIndex = ceil(progression * (positions.size - 1)).toInt()
-                if (!positions.indices.contains(positionIndex)) {
-                    return
-                }
-
-                val locator = positions[positionIndex].copyWithLocations(progression = progression)
-                // Pageindex is actually the page number so to get a zero based index we subtract one.
-                paginationListener.onPageChanged(pageIndex - 1, totalPages, locator)
-            }
         }
 
         override fun onPageEnded(end: Boolean) {
@@ -466,16 +454,34 @@ class EpubNavigatorFragment private constructor(
             }
         }
 
-        override fun onTap(point: PointF): Boolean {
-            point.adjustToViewport()
-            return listener?.onTap(point) ?: false
-        }
+        override fun onTap(point: PointF): Boolean =
+            listener?.onTap(point.adjustedToViewport()) ?: false
 
-        override fun onDecorationActivated(id: DecorationId, group: String, rect: RectF, point: PointF): Boolean {
-            rect.adjustToViewport()
-            point.adjustToViewport()
-            return viewModel.onDecorationActivated(id, group, rect, point)
-        }
+        override fun onDragStart(event: R2BasicWebView.DragEvent): Boolean =
+            listener?.onDragStart(
+                startPoint = event.startPoint.adjustedToViewport(),
+                offset = event.offset
+            ) ?: false
+
+        override fun onDragMove(event: R2BasicWebView.DragEvent): Boolean =
+            listener?.onDragMove(
+                startPoint = event.startPoint.adjustedToViewport(),
+                offset = event.offset
+            ) ?: false
+
+        override fun onDragEnd(event: R2BasicWebView.DragEvent): Boolean =
+            listener?.onDragEnd(
+                startPoint = event.startPoint.adjustedToViewport(),
+                offset = event.offset
+            ) ?: false
+
+        override fun onDecorationActivated(id: DecorationId, group: String, rect: RectF, point: PointF): Boolean =
+            viewModel.onDecorationActivated(
+                id = id,
+                group = group,
+                rect = rect.adjustedToViewport(),
+                point = point.adjustedToViewport()
+            )
 
         override fun onProgressionChanged() {
             notifyCurrentLocation()
@@ -691,16 +697,17 @@ class EpubNavigatorFragment private constructor(
         debounceLocationNotificationJob = launch {
             delay(100L)
 
-            if (pendingLocator != null) {
+            val webView = currentFragment?.webView
+            if (pendingLocator != null || webView == null) {
                 return@launch
             }
 
             // The transition has stabilized, so we can ask the web view to refresh its current
             // item to reflect the current scroll position.
-            currentFragment?.webView?.updateCurrentItem()
+            webView.updateCurrentItem()
 
             val resource = publication.readingOrder[resourcePager.currentItem]
-            val progression = currentFragment?.webView?.progression?.coerceIn(0.0, 1.0) ?: 0.0
+            val progression = webView.progression.coerceIn(0.0, 1.0)
             val positions = publication.positionsByResource[resource.href]?.takeIf { it.isNotEmpty() }
                     ?: return@launch
 
@@ -713,12 +720,16 @@ class EpubNavigatorFragment private constructor(
                     .copy(title = tableOfContentsTitleByHref[resource.href])
                     .copyWithLocations(progression = progression)
 
-            if (locator == _currentLocator.value) {
-                return@launch
-            }
-
             _currentLocator.value = locator
+
+            // Deprecated notifications
+
             navigatorDelegate?.locationDidChange(navigator = navigator, locator = locator)
+            paginationListener?.onPageChanged(
+                pageIndex = webView.mCurItem,
+                totalPages = webView.numPages,
+                locator = locator
+            )
         }
     }